Revert "[CARBONDATA-2532][Integration] Carbon to support spark 2.3 version, compatability issues"
This reverts commit d0fa52396687ccc1a5d029006e7204771c04a9eb.
