+++ integration/spark-common-test/src/test/scala/org/apache/carbondata/spark/testsuite/createTable/TestCreateTableUsingSparkCarbonFileFormat.scala
@@ -21,7 +21,6 @@ import java.io.File
 
 import org.apache.commons.io.FileUtils
 import org.apache.spark.sql.test.util.QueryTest
-import org.apache.spark.util.SparkUtil
 import org.scalatest.BeforeAndAfterAll
 
 import org.apache.carbondata.common.exceptions.sql.MalformedCarbonCommandException
@@ -112,10 +111,10 @@ class TestCreateTableUsingSparkCarbonFileFormat extends QueryTest with BeforeAnd
     sql("DROP TABLE IF EXISTS sdkOutputTable")
 
     //data source file format
-    if (SparkUtil.isSparkVersionEqualToX("2.1")) {
+    if (sqlContext.sparkContext.version.startsWith("2.1")) {
       //data source file format
       sql(s"""CREATE TABLE sdkOutputTable USING carbonfile OPTIONS (PATH '$filePath') """)
-    } else if (SparkUtil.isSparkVersionXandAbove("2.2")) {
+    } else if (sqlContext.sparkContext.version.startsWith("2.2")) {
       //data source file format
       sql(
         s"""CREATE TABLE sdkOutputTable USING carbonfile LOCATION
@@ -160,10 +159,10 @@ class TestCreateTableUsingSparkCarbonFileFormat extends QueryTest with BeforeAnd
     sql("DROP TABLE IF EXISTS sdkOutputTable")
 
     //data source file format
-    if (SparkUtil.isSparkVersionEqualToX("2.1")) {
+    if (sqlContext.sparkContext.version.startsWith("2.1")) {
       //data source file format
       sql(s"""CREATE TABLE sdkOutputTable USING carbonfile OPTIONS (PATH '$filePath') """)
-    } else if (SparkUtil.isSparkVersionXandAbove("2.2")) {
+    } else if (sqlContext.sparkContext.version.startsWith("2.2")) {
       //data source file format
       sql(
         s"""CREATE TABLE sdkOutputTable USING carbonfile LOCATION
@@ -189,10 +188,10 @@ class TestCreateTableUsingSparkCarbonFileFormat extends QueryTest with BeforeAnd
     sql("DROP TABLE IF EXISTS sdkOutputTable")
 
 
-    if (SparkUtil.isSparkVersionEqualToX("2.1")) {
+    if (sqlContext.sparkContext.version.startsWith("2.1")) {
       //data source file format
       sql(s"""CREATE TABLE sdkOutputTable USING carbonfile OPTIONS (PATH '$filePath') """)
-    } else if (SparkUtil.isSparkVersionXandAbove("2.2")) {
+    } else if (sqlContext.sparkContext.version.startsWith("2.2")) {
       //data source file format
       sql(
         s"""CREATE TABLE sdkOutputTable USING carbonfile LOCATION
@@ -222,10 +221,10 @@ class TestCreateTableUsingSparkCarbonFileFormat extends QueryTest with BeforeAnd
 
     val exception = intercept[org.apache.spark.SparkException] {
       //    data source file format
-      if (SparkUtil.isSparkVersionEqualToX("2.1")) {
+      if (sqlContext.sparkContext.version.startsWith("2.1")) {
         //data source file format
         sql(s"""CREATE TABLE sdkOutputTable USING carbonfile OPTIONS (PATH '$filePath') """)
-      } else if (SparkUtil.isSparkVersionXandAbove("2.2")) {
+      } else if (sqlContext.sparkContext.version.startsWith("2.2")) {
         //data source file format
         sql(
           s"""CREATE TABLE sdkOutputTable USING carbonfile LOCATION
@@ -252,10 +251,10 @@ class TestCreateTableUsingSparkCarbonFileFormat extends QueryTest with BeforeAnd
 
     val exception = intercept[org.apache.spark.SparkException] {
       //data source file format
-      if (SparkUtil.isSparkVersionEqualToX("2.1")) {
+      if (sqlContext.sparkContext.version.startsWith("2.1")) {
         //data source file format
         sql(s"""CREATE TABLE sdkOutputTable USING carbonfile OPTIONS (PATH '$filePath') """)
-      } else if (SparkUtil.isSparkVersionXandAbove("2.2")) {
+      } else if (sqlContext.sparkContext.version.startsWith("2.2")) {
         //data source file format
         sql(
           s"""CREATE TABLE sdkOutputTable USING carbonfile LOCATION
@@ -282,10 +281,10 @@ class TestCreateTableUsingSparkCarbonFileFormat extends QueryTest with BeforeAnd
     //data source file format
     sql("DROP TABLE IF EXISTS sdkOutputTable")
 
-    if (SparkUtil.isSparkVersionEqualToX("2.1")) {
+    if (sqlContext.sparkContext.version.startsWith("2.1")) {
       //data source file format
       sql(s"""CREATE TABLE sdkOutputTable USING carbonfile OPTIONS (PATH '$filePath') """)
-    } else if (SparkUtil.isSparkVersionXandAbove("2.2")) {
+    } else if (sqlContext.sparkContext.version.startsWith("2.2")) {
       //data source file format
       sql(
         s"""CREATE TABLE sdkOutputTable USING carbonfile LOCATION
@@ -331,10 +330,10 @@ class TestCreateTableUsingSparkCarbonFileFormat extends QueryTest with BeforeAnd
     assert(new File(filePath).exists())
     sql("DROP TABLE IF EXISTS sdkOutputTable")
 
-    if (SparkUtil.isSparkVersionEqualToX("2.1")) {
+    if (sqlContext.sparkContext.version.startsWith("2.1")) {
       //data source file format
       sql(s"""CREATE TABLE sdkOutputTable USING carbonfile OPTIONS (PATH '$filePath') """)
-    } else if (SparkUtil.isSparkVersionXandAbove("2.2")) {
+    } else if (sqlContext.sparkContext.version.startsWith("2.2")) {
       //data source file format
       sql(
         s"""CREATE TABLE sdkOutputTable USING carbonfile LOCATION
