+++ integration/spark-common-test/src/test/scala/org/apache/carbondata/spark/testsuite/filterexpr/AllDataTypesTestCaseFilter.scala
@@ -17,7 +17,6 @@
 
 package org.apache.carbondata.spark.testsuite.filterexpr
 
-import org.apache.spark.sql.execution.BatchedDataSourceScanExec
 import org.apache.spark.sql.test.util.QueryTest
 import org.scalatest.BeforeAndAfterAll
 
@@ -57,15 +56,13 @@ class AllDataTypesTestCaseFilter extends QueryTest with BeforeAndAfterAll {
   test("verify like query ends with filter push down") {
     val df = sql("select * from alldatatypestableFilter where empname like '%nandh'").queryExecution
       .sparkPlan
-    assert(df.asInstanceOf[BatchedDataSourceScanExec].metadata
-      .get("PushedFilters").get.contains("CarbonEndsWith"))
+    assert(df.metadata.get("PushedFilters").get.contains("CarbonEndsWith"))
   }
 
   test("verify like query contains with filter push down") {
     val df = sql("select * from alldatatypestableFilter where empname like '%nand%'").queryExecution
       .sparkPlan
-    assert(df.asInstanceOf[BatchedDataSourceScanExec].metadata
-      .get("PushedFilters").get.contains("CarbonContainsWith"))
+    assert(df.metadata.get("PushedFilters").get.contains("CarbonContainsWith"))
   }
   
   override def afterAll {
