+++ integration/spark-common-test/src/test/scala/org/apache/carbondata/sql/commands/StoredAsCarbondataSuite.scala
@@ -87,8 +87,7 @@ class StoredAsCarbondataSuite extends QueryTest with BeforeAndAfterEach {
       sql("CREATE TABLE carbon_table(key INT, value STRING) STORED AS  ")
     } catch {
       case e: Exception =>
-        assert(e.getMessage.contains("no viable alternative at input") ||
-        e.getMessage.contains("mismatched input '<EOF>' expecting "))
+        assert(e.getMessage.contains("no viable alternative at input"))
     }
   }
