+++ integration/spark2/src/main/scala/org/apache/spark/sql/CarbonCountStar.scala
@@ -30,8 +30,6 @@ import org.apache.spark.sql.catalyst.{InternalRow, TableIdentifier}
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.execution.LeafExecNode
 import org.apache.spark.sql.optimizer.CarbonFilters
-import org.apache.spark.sql.types.StringType
-import org.apache.spark.unsafe.types.UTF8String
 
 import org.apache.carbondata.core.metadata.AbsoluteTableIdentifier
 import org.apache.carbondata.core.metadata.schema.table.CarbonTable
@@ -62,13 +60,7 @@ case class CarbonCountStar(
             carbonTable.getTableName,
             Some(carbonTable.getDatabaseName))).map(_.asJava).orNull),
       carbonTable)
-    val valueRaw =
-      attributesRaw.head.dataType match {
-        case StringType => Seq(UTF8String.fromString(Long.box(rowCount).toString)).toArray
-          .asInstanceOf[Array[Any]]
-        case _ => Seq(Long.box(rowCount)).toArray.asInstanceOf[Array[Any]]
-      }
-    val value = new GenericInternalRow(valueRaw)
+    val value = new GenericInternalRow(Seq(Long.box(rowCount)).toArray.asInstanceOf[Array[Any]])
     val unsafeProjection = UnsafeProjection.create(output.map(_.dataType).toArray)
     val row = if (outUnsafeRows) unsafeProjection(value) else value
     sparkContext.parallelize(Seq(row))
