+++ integration/spark2/src/main/scala/org/apache/spark/sql/execution/command/management/CarbonLoadDataCommand.scala
@@ -829,21 +829,17 @@ case class CarbonLoadDataCommand(
       // datatype is always int
       val column = table.getColumnByName(table.getTableName, attr.name)
       if (column.hasEncoding(Encoding.DICTIONARY)) {
-        CarbonToSparkAdapater.createAttributeReference(attr.name,
+        AttributeReference(
+          attr.name,
           IntegerType,
           attr.nullable,
-          attr.metadata,
-          attr.exprId,
-          attr.qualifier,
-          attr)
+          attr.metadata)(attr.exprId, attr.qualifier, attr.isGenerated)
       } else if (attr.dataType == TimestampType || attr.dataType == DateType) {
-        CarbonToSparkAdapater.createAttributeReference(attr.name,
+        AttributeReference(
+          attr.name,
           LongType,
           attr.nullable,
-          attr.metadata,
-          attr.exprId,
-          attr.qualifier,
-          attr)
+          attr.metadata)(attr.exprId, attr.qualifier, attr.isGenerated)
       } else {
         attr
       }
@@ -1041,8 +1037,7 @@ case class CarbonLoadDataCommand(
 
     CarbonReflectionUtils.getLogicalRelation(hdfsRelation,
       hdfsRelation.schema.toAttributes,
-      Some(catalogTable),
-      false)
+      Some(catalogTable))
   }
