+++ integration/spark2/src/main/scala/org/apache/spark/sql/hive/CarbonPreAggregateRules.scala
@@ -273,14 +273,11 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
         case attr: AttributeReference =>
           updatedExpression.find { p => p._1.sameRef(attr) } match {
             case Some((_, childAttr)) =>
-              CarbonToSparkAdapater.createAttributeReference(
+              AttributeReference(
                 childAttr.name,
                 childAttr.dataType,
                 childAttr.nullable,
-                childAttr.metadata,
-                childAttr.exprId,
-                attr.qualifier,
-                attr)
+                childAttr.metadata)(childAttr.exprId, attr.qualifier, attr.isGenerated)
             case None =>
               attr
           }
@@ -299,14 +296,11 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
         case attr: AttributeReference =>
           updatedExpression.find { p => p._1.sameRef(attr) } match {
             case Some((_, childAttr)) =>
-              CarbonToSparkAdapater.createAttributeReference(
+              AttributeReference(
                 childAttr.name,
                 childAttr.dataType,
                 childAttr.nullable,
-                childAttr.metadata,
-                childAttr.exprId,
-                attr.qualifier,
-                attr)
+                childAttr.metadata)(childAttr.exprId, attr.qualifier, attr.isGenerated)
             case None =>
               attr
           }
@@ -783,36 +777,24 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
         val factAlias = factPlanExpForStreaming(name)
         // create attribute reference object for each expression
         val attrs = factAlias.map { factAlias =>
-          CarbonToSparkAdapater.createAttributeReference(
+          AttributeReference(
             name,
             alias.dataType,
-            alias.nullable,
-            Metadata.empty,
-            factAlias.exprId,
-            alias.qualifier,
-            alias)
+            alias.nullable) (factAlias.exprId, alias.qualifier, alias.isGenerated)
         }
         // add aggregate function in Aggregate node added for handling streaming
         // to aggregate results from fact and aggregate table
         val updatedAggExp = getAggregateExpressionForAggregation(aggExp, attrs)
         // same reference id will be used as it can be used by above nodes in the plan like
         // sort, project, join
-        CarbonToSparkAdapater.createAliasRef(
+        Alias(
           updatedAggExp.head,
-          name,
-          alias.exprId,
-          alias.qualifier,
-          Option(alias.metadata),
-          Some(alias))
+          name)(alias.exprId, alias.qualifier, Option(alias.metadata), alias.isGenerated)
       case alias@Alias(expression, name) =>
-        CarbonToSparkAdapater.createAttributeReference(
+        AttributeReference(
           name,
           alias.dataType,
-          alias.nullable,
-          Metadata.empty,
-          alias.exprId,
-          alias.qualifier,
-          alias)
+          alias.nullable) (alias.exprId, alias.qualifier, alias.isGenerated)
     }
     updatedExp
   }
@@ -915,10 +897,9 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
           case attr: AttributeReference =>
             newAggExp += attr
           case exp: Expression =>
-            newAggExp += CarbonToSparkAdapater.createAliasRef(
+            newAggExp += Alias(
               exp,
-              "dummy_" + counter,
-              NamedExpression.newExprId)
+              "dummy_" + counter)(NamedExpression.newExprId, None, None, false)
             counter = counter + 1
         }
       }
@@ -942,12 +923,12 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
         // get the new aggregate expression
         val newAggExp = getAggFunctionForFactStreaming(aggExp)
         val updatedExp = newAggExp.map { exp =>
-          CarbonToSparkAdapater.createAliasRef(exp,
-            name,
-            NamedExpression.newExprId,
-            alias.qualifier,
+          Alias(exp,
+              name)(
+              NamedExpression.newExprId,
+              alias.qualifier,
             Some(alias.metadata),
-            Some(alias))
+              alias.isGenerated)
         }
         // adding to map which will be used while Adding an Aggregate node for handling streaming
         // table plan change
@@ -955,13 +936,10 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
         updatedExp
       case alias@Alias(exp: Expression, name) =>
         val newAlias = Seq(alias)
-        val attr = CarbonToSparkAdapater.createAttributeReference(name,
-          alias.dataType,
-          alias.nullable,
-          alias.metadata,
-          alias.exprId,
-          alias.qualifier,
-          alias)
+        val attr = AttributeReference(name,
+            alias.dataType,
+            alias.nullable,
+            alias.metadata) (alias.exprId, alias.qualifier, alias.isGenerated)
         factPlanGrpExpForStreaming.put(
           AggExpToColumnMappingModel(
             removeQualifiers(PreAggregateUtil.normalizeExprId(exp, plan.allAttributes))),
@@ -1115,14 +1093,11 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
   private def removeQualifiers(expression: Expression) : Expression = {
     expression.transform {
       case attr: AttributeReference =>
-        CarbonToSparkAdapater.createAttributeReference(
+        AttributeReference(
           attr.name,
           attr.dataType,
           attr.nullable,
-          attr.metadata,
-          attr.exprId,
-          None,
-          attr)
+          attr.metadata)(attr.exprId, None, attr.isGenerated)
     }
   }
 
@@ -1388,13 +1363,10 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
           attr,
           attributes)
         val newExpressionId = NamedExpression.newExprId
-        val childTableAttr = CarbonToSparkAdapater.createAttributeReference(attr.name,
+        val childTableAttr = AttributeReference(attr.name,
           childAttr.dataType,
           childAttr.nullable,
-          childAttr.metadata,
-          newExpressionId,
-          childAttr.qualifier,
-          attr)
+          childAttr.metadata)(newExpressionId, childAttr.qualifier, attr.isGenerated)
         updatedExpression.put(attr, childTableAttr)
         // returning the alias to show proper column name in output
         Seq(Alias(childAttr,
@@ -1406,20 +1378,12 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
           attr,
           attributes)
         val newExpressionId = NamedExpression.newExprId
-        val parentTableAttr = CarbonToSparkAdapater.createAttributeReference(name,
+        val parentTableAttr = AttributeReference(name,
           alias.dataType,
-          alias.nullable,
-          Metadata.empty,
-          alias.exprId,
-          alias.qualifier,
-          alias)
-        val childTableAttr = CarbonToSparkAdapater.createAttributeReference(name,
+          alias.nullable) (alias.exprId, alias.qualifier, alias.isGenerated)
+        val childTableAttr = AttributeReference(name,
           alias.dataType,
-          alias.nullable,
-          Metadata.empty,
-          newExpressionId,
-          alias.qualifier,
-          alias)
+          alias.nullable) (newExpressionId, alias.qualifier, alias.isGenerated)
         updatedExpression.put(parentTableAttr, childTableAttr)
         // returning alias with child attribute reference
         Seq(Alias(childAttr,
@@ -1445,21 +1409,13 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
           val newExpressionId = NamedExpression.newExprId
           // create a parent attribute reference which will be replced on node which may be referred
           // by node like sort join
-          val parentTableAttr = CarbonToSparkAdapater.createAttributeReference(name,
+          val parentTableAttr = AttributeReference(name,
             alias.dataType,
-            alias.nullable,
-            Metadata.empty,
-            alias.exprId,
-            alias.qualifier,
-            alias)
+            alias.nullable)(alias.exprId, alias.qualifier, alias.isGenerated)
           // creating a child attribute reference which will be replced
-          val childTableAttr = CarbonToSparkAdapater.createAttributeReference(name,
+          val childTableAttr = AttributeReference(name,
             alias.dataType,
-            alias.nullable,
-            Metadata.empty,
-            newExpressionId,
-            alias.qualifier,
-            alias)
+            alias.nullable)(newExpressionId, alias.qualifier, alias.isGenerated)
           // adding to map, will be used during other node updation like sort, join, project
           updatedExpression.put(parentTableAttr, childTableAttr)
           // returning alias with child attribute reference
@@ -1470,12 +1426,12 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
           // for streaming table
           // create alias for aggregate table
           val aggExpForStreaming = aggExp.map{ exp =>
-            CarbonToSparkAdapater.createAliasRef(exp,
-              name,
+            Alias(exp,
+              name)(
               NamedExpression.newExprId,
               alias.qualifier,
               Some(alias.metadata),
-              Some(alias)).asInstanceOf[NamedExpression]
+              alias.isGenerated).asInstanceOf[NamedExpression]
           }
           aggExpForStreaming
         }
@@ -1504,20 +1460,12 @@ case class CarbonPreAggregateQueryRules(sparkSession: SparkSession) extends Rule
             }
           }
         val newExpressionId = NamedExpression.newExprId
-        val parentTableAttr = CarbonToSparkAdapater.createAttributeReference(name,
+        val parentTableAttr = AttributeReference(name,
           alias.dataType,
-          alias.nullable,
-          Metadata.empty,
-          alias.exprId,
-          alias.qualifier,
-          alias)
-        val childTableAttr = CarbonToSparkAdapater.createAttributeReference(name,
+          alias.nullable) (alias.exprId, alias.qualifier, alias.isGenerated)
+        val childTableAttr = AttributeReference(name,
           alias.dataType,
-          alias.nullable,
-          Metadata.empty,
-          newExpressionId,
-          alias.qualifier,
-          alias)
+          alias.nullable) (newExpressionId, alias.qualifier, alias.isGenerated)
         updatedExpression.put(parentTableAttr, childTableAttr)
         Seq(Alias(updatedExp, name)(newExpressionId,
           alias.qualifier).asInstanceOf[NamedExpression])
@@ -1839,23 +1787,20 @@ case class CarbonPreAggregateDataLoadingRules(sparkSession: SparkSession)
               // named expression list otherwise update the list and add it to set
               if (!validExpressionsMap.contains(AggExpToColumnMappingModel(sumExp))) {
                 namedExpressionList +=
-                CarbonToSparkAdapater.createAliasRef(expressions.head,
-                  name + "_ sum",
-                  NamedExpression.newExprId,
+                Alias(expressions.head, name + "_ sum")(NamedExpression.newExprId,
                   alias.qualifier,
                   Some(alias.metadata),
-                  Some(alias))
+                  alias.isGenerated)
                 validExpressionsMap += AggExpToColumnMappingModel(sumExp)
               }
               // check with same expression already count is present then do not add to
               // named expression list otherwise update the list and add it to set
               if (!validExpressionsMap.contains(AggExpToColumnMappingModel(countExp))) {
                 namedExpressionList +=
-                CarbonToSparkAdapater.createAliasRef(expressions.last, name + "_ count",
-                    NamedExpression.newExprId,
-                    alias.qualifier,
-                    Some(alias.metadata),
-                    Some(alias))
+                Alias(expressions.last, name + "_ count")(NamedExpression.newExprId,
+                  alias.qualifier,
+                  Some(alias.metadata),
+                  alias.isGenerated)
                 validExpressionsMap += AggExpToColumnMappingModel(countExp)
               }
             } else {
