+++ integration/spark2/src/main/scala/org/apache/spark/sql/optimizer/CarbonFilters.scala
@@ -31,7 +31,6 @@ import org.apache.spark.sql.CarbonEndsWith
 import org.apache.spark.sql.CarbonExpressions.{MatchCast => Cast}
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.hive.CarbonSessionCatalog
-import org.apache.spark.util.{CarbonReflectionUtils, SparkUtil}
 
 import org.apache.carbondata.core.constants.CarbonCommonConstants
 import org.apache.carbondata.core.datamap.Segment
@@ -48,7 +47,6 @@ import org.apache.carbondata.core.util.ThreadLocalSessionInfo
 import org.apache.carbondata.datamap.{TextMatch, TextMatchLimit}
 import org.apache.carbondata.spark.CarbonAliasDecoderRelation
 import org.apache.carbondata.spark.util.CarbonScalaUtil
-
 
 
 /**
@@ -296,27 +294,6 @@ object CarbonFilters {
     filters.flatMap(translate(_, false)).toArray
   }
 
-  /**
-   * This API checks whether StringTrim object is compatible with
-   * carbon,carbon only deals with the space any other symbol should
-   * be ignored.So condition is SPARK version < 2.3.
-   * If it is 2.3 then trimStr field should be empty
-   *
-   * @param stringTrim
-   * @return
-   */
-  def isStringTrimCompatibleWithCarbon(stringTrim: StringTrim): Boolean = {
-    var isCompatible = true
-    if (SparkUtil.isSparkVersionXandAbove("2.3")) {
-      val trimStr = CarbonReflectionUtils.getField("trimStr", stringTrim)
-        .asInstanceOf[Option[Expression]]
-      if (trimStr.isDefined) {
-        isCompatible = false
-      }
-    }
-    isCompatible
-  }
-
   def transformExpression(expr: Expression): CarbonExpression = {
     expr match {
       case Or(left, right)
@@ -404,8 +381,7 @@ object CarbonFilters {
           new CarbonLiteralExpression(maxValueLimit,
             CarbonScalaUtil.convertSparkToCarbonDataType(dataType)))
         new AndExpression(l, r)
-      case strTrim: StringTrim if isStringTrimCompatibleWithCarbon(strTrim) =>
-        transformExpression(strTrim)
+      case StringTrim(child) => transformExpression(child)
       case s: ScalaUDF =>
         new MatchExpression(s.children.head.toString())
       case _ =>
