+++ integration/spark2/src/main/spark2.2/org/apache/spark/sql/hive/CarbonSessionState.scala
@@ -18,25 +18,30 @@ package org.apache.spark.sql.hive
 
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.Path
+import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.catalyst.analysis.{Analyzer, FunctionRegistry}
 import org.apache.spark.sql.catalyst.catalog._
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.catalyst.optimizer.Optimizer
 import org.apache.spark.sql.catalyst.parser.ParserInterface
-import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
+import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan}
+import org.apache.spark.sql.catalyst.parser.{ParserInterface, SqlBaseParser}
+import org.apache.spark.sql.catalyst.parser.ParserUtils.{string, withOrigin}
+import org.apache.spark.sql.catalyst.parser.SqlBaseParser.{AddTableColumnsContext, ChangeColumnContext, CreateHiveTableContext, CreateTableContext, ShowTablesContext}
+import org.apache.spark.sql.catalyst.plans.logical.{Filter, LogicalPlan, SubqueryAlias}
 import org.apache.spark.sql.catalyst.rules.Rule
-import org.apache.spark.sql.execution.datasources._
+import org.apache.spark.sql.execution.datasources.{FindDataSourceTable, PreWriteCheck, ResolveSQLOnFile, _}
+import org.apache.spark.sql.execution.command._
+import org.apache.spark.sql.execution.command.schema.{CarbonAlterTableAddColumnCommand, CarbonAlterTableDataTypeChangeCommand}
+import org.apache.spark.sql.execution.command.table.{CarbonExplainCommand, CarbonShowTablesCommand}
+import org.apache.spark.sql.execution.datasources.{FindDataSourceTable, LogicalRelation, PreWriteCheck, ResolveSQLOnFile, _}
 import org.apache.spark.sql.execution.strategy.{CarbonLateDecodeStrategy, DDLStrategy, StreamingTableStrategy}
-import org.apache.spark.sql.internal.{SQLConf, SessionResourceLoader, SessionState, SessionStateBuilder}
+import org.apache.spark.sql.hive.client.HiveClient
+import org.apache.spark.sql.internal.{SQLConf, SessionState}
 import org.apache.spark.sql.optimizer.{CarbonIUDRule, CarbonLateDecodeRule, CarbonUDFTransformRule}
 import org.apache.spark.sql.parser.CarbonSparkSqlParser
-import org.apache.spark.sql.types.{StructField, StructType}
-import org.apache.spark.sql.{CarbonEnv, SparkSession}
 
-import org.apache.carbondata.core.util.CarbonUtil
-import org.apache.carbondata.core.util.path.CarbonTablePath
-import org.apache.carbondata.format.TableInfo
 import org.apache.carbondata.spark.util.CarbonScalaUtil
 
 /**
@@ -51,8 +56,8 @@ import org.apache.carbondata.spark.util.CarbonScalaUtil
  * @param conf
  * @param hadoopConf
  */
-class InMemorySessionCatalog(
-    externalCatalog: ExternalCatalog,
+class CarbonHiveSessionCatalog(
+    externalCatalog: HiveExternalCatalog,
     globalTempViewManager: GlobalTempViewManager,
     functionRegistry: FunctionRegistry,
     sparkSession: SparkSession,
@@ -60,9 +65,10 @@ class InMemorySessionCatalog(
     hadoopConf: Configuration,
     parser: ParserInterface,
     functionResourceLoader: FunctionResourceLoader)
-  extends SessionCatalog(
+  extends HiveSessionCatalog (
     externalCatalog,
     globalTempViewManager,
+    new HiveMetastoreCatalog(sparkSession),
     functionRegistry,
     conf,
     hadoopConf,
@@ -70,89 +76,21 @@ class InMemorySessionCatalog(
     functionResourceLoader
   ) with CarbonSessionCatalog {
 
-  override def alterTableRename(oldTableIdentifier: TableIdentifier,
-      newTableIdentifier: TableIdentifier,
-      newTablePath: String): Unit = {
-    sparkSession.sessionState.catalog.renameTable(oldTableIdentifier, newTableIdentifier)
-  }
-
-  override def alterTable(tableIdentifier: TableIdentifier,
-      schemaParts: String,
-      cols: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
-  : Unit = {
-    // NOt Required in case of In-memory catalog
-  }
-
-  override def alterAddColumns(tableIdentifier: TableIdentifier,
-      schemaParts: String,
-      newColumns: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
-  : Unit = {
-    val catalogTable = sparkSession.sessionState.catalog.getTableMetadata(tableIdentifier)
-    val structType = catalogTable.schema
-    var newStructType = structType
-    newColumns.get.foreach {cols =>
-      newStructType = structType
-        .add(cols.getColumnName, CarbonScalaUtil.convertCarbonToSparkDataType(cols.getDataType))
-    }
-    alterSchema(newStructType, catalogTable, tableIdentifier)
-  }
-
-  override def alterDropColumns(tableIdentifier: TableIdentifier,
-      schemaParts: String,
-      dropCols: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
-  : Unit = {
-    val catalogTable = sparkSession.sessionState.catalog.getTableMetadata(tableIdentifier)
-    val fields = catalogTable.schema.fields.filterNot { field =>
-      dropCols.get.exists { col =>
-        col.getColumnName.equalsIgnoreCase(field.name)
-      }
-    }
-    alterSchema(new StructType(fields), catalogTable, tableIdentifier)
-  }
-
-  override def alterColumnChangeDataType(tableIdentifier: TableIdentifier,
-      schemaParts: String,
-      columns: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
-  : Unit = {
-    val catalogTable = sparkSession.sessionState.catalog.getTableMetadata(tableIdentifier)
-    val a = catalogTable.schema.fields.flatMap { field =>
-      columns.get.map { col =>
-        if (col.getColumnName.equalsIgnoreCase(field.name)) {
-          StructField(col.getColumnName,
-            CarbonScalaUtil.convertCarbonToSparkDataType(col.getDataType))
-        } else {
-          field
-        }
-      }
-    }
-    alterSchema(new StructType(a), catalogTable, tableIdentifier)
-  }
-
-  private def alterSchema(structType: StructType,
-      catalogTable: CatalogTable,
-      tableIdentifier: TableIdentifier): Unit = {
-    val copy = catalogTable.copy(schema = structType)
-    sparkSession.sessionState.catalog.alterTable(copy)
-    sparkSession.sessionState.catalog.refreshTable(tableIdentifier)
-  }
-
-  lazy val carbonEnv = {
+  private lazy val carbonEnv = {
     val env = new CarbonEnv
     env.init(sparkSession)
     env
   }
-
-  def getCarbonEnv() : CarbonEnv = {
+  /**
+   * return's the carbonEnv instance
+   * @return
+   */
+  override def getCarbonEnv() : CarbonEnv = {
     carbonEnv
   }
 
   // Initialize all listeners to the Operation bus.
   CarbonEnv.initListeners()
-
-  def getThriftTableInfo(tablePath: String): TableInfo = {
-    val tableMetadataFile = CarbonTablePath.getSchemaFilePath(tablePath)
-    CarbonUtil.readSchemaFile(tableMetadataFile)
-  }
 
   override def lookupRelation(name: TableIdentifier): LogicalPlan = {
     val rtnRelation = super.lookupRelation(name)
@@ -170,8 +108,52 @@ class InMemorySessionCatalog(
    *
    * @return
    */
-  def getClient(): org.apache.spark.sql.hive.client.HiveClient = {
-    null
+  override def getClient(): org.apache.spark.sql.hive.client.HiveClient = {
+    sparkSession.asInstanceOf[CarbonSession].sharedState.externalCatalog
+      .asInstanceOf[HiveExternalCatalog].client
+  }
+
+  def alterTableRename(oldTableIdentifier: TableIdentifier,
+      newTableIdentifier: TableIdentifier,
+      newTablePath: String): Unit = {
+    getClient().runSqlHive(
+      s"ALTER TABLE ${ oldTableIdentifier.database.get }.${ oldTableIdentifier.table } " +
+      s"RENAME TO ${ oldTableIdentifier.database.get }.${ newTableIdentifier.table }")
+    getClient().runSqlHive(
+      s"ALTER TABLE ${ oldTableIdentifier.database.get }.${ newTableIdentifier.table} " +
+      s"SET SERDEPROPERTIES" +
+      s"('tableName'='${ newTableIdentifier.table }', " +
+      s"'dbName'='${ oldTableIdentifier.database.get }', 'tablePath'='${ newTablePath }')")
+  }
+
+  override def alterTable(tableIdentifier: TableIdentifier,
+      schemaParts: String,
+      cols: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
+  : Unit = {
+    getClient()
+      .runSqlHive(s"ALTER TABLE ${tableIdentifier.database.get}.${ tableIdentifier.table } " +
+                  s"SET TBLPROPERTIES(${ schemaParts })")
+  }
+
+  override def alterAddColumns(tableIdentifier: TableIdentifier,
+      schemaParts: String,
+      cols: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
+  : Unit = {
+    alterTable(tableIdentifier, schemaParts, cols)
+  }
+
+  override def alterDropColumns(tableIdentifier: TableIdentifier,
+      schemaParts: String,
+      cols: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
+  : Unit = {
+    alterTable(tableIdentifier, schemaParts, cols)
+  }
+
+  override def alterColumnChangeDataType(tableIdentifier: TableIdentifier,
+      schemaParts: String,
+      cols: Option[Seq[org.apache.carbondata.core.metadata.schema.table.column.ColumnSchema]])
+  : Unit = {
+    alterTable(tableIdentifier, schemaParts, cols)
   }
 
   override def createPartitions(
@@ -214,16 +196,21 @@ class InMemorySessionCatalog(
   }
 }
 
-class CarbonInMemorySessionStateBuilder (sparkSession: SparkSession,
+/**
+ * Session state implementation to override sql parser and adding strategies
+ *
+ * @param sparkSession
+ */
+class CarbonSessionStateBuilder(sparkSession: SparkSession,
     parentState: Option[SessionState] = None)
-  extends SessionStateBuilder(sparkSession, parentState) {
+  extends HiveSessionStateBuilder(sparkSession, parentState) {
 
   override lazy val sqlParser: ParserInterface = new CarbonSparkSqlParser(conf, sparkSession)
 
   experimentalMethods.extraStrategies =
     Seq(new StreamingTableStrategy(sparkSession),
-      new CarbonLateDecodeStrategy,
-      new DDLStrategy(sparkSession)
+        new CarbonLateDecodeStrategy,
+        new DDLStrategy(sparkSession)
     )
   experimentalMethods.extraOptimizations = Seq(new CarbonIUDRule,
     new CarbonUDFTransformRule,
@@ -232,8 +219,11 @@ class CarbonInMemorySessionStateBuilder (sparkSession: SparkSession,
   /**
    * Internal catalog for managing table and database states.
    */
-  override protected lazy val catalog: InMemorySessionCatalog = {
-    val catalog = new InMemorySessionCatalog(
+  /**
+   * Create a [[CarbonSessionStateBuilder]].
+   */
+  override protected lazy val catalog: CarbonHiveSessionCatalog = {
+    val catalog = new CarbonHiveSessionCatalog(
       externalCatalog,
       session.sharedState.globalTempViewManager,
       functionRegistry,
@@ -246,31 +236,42 @@ class CarbonInMemorySessionStateBuilder (sparkSession: SparkSession,
     catalog
   }
 
-  private def externalCatalog: ExternalCatalog =
-    session.sharedState.externalCatalog.asInstanceOf[ExternalCatalog]
+  private def externalCatalog: HiveExternalCatalog =
+    session.sharedState.externalCatalog.asInstanceOf[HiveExternalCatalog]
 
-  override protected lazy val resourceLoader: SessionResourceLoader = {
-    new SessionResourceLoader(session)
+  /**
+   * Create a Hive aware resource loader.
+   */
+  override protected lazy val resourceLoader: HiveSessionResourceLoader = {
+    val client: HiveClient = externalCatalog.client.newSession()
+    new HiveSessionResourceLoader(session, client)
   }
 
   override lazy val optimizer: Optimizer = new CarbonOptimizer(catalog, conf, experimentalMethods)
 
   override protected def analyzer: Analyzer = new CarbonAnalyzer(catalog, conf, sparkSession,
     new Analyzer(catalog, conf) {
+
       override val extendedResolutionRules: Seq[Rule[LogicalPlan]] =
+        new ResolveHiveSerdeTable(session) +:
         new FindDataSourceTable(session) +:
         new ResolveSQLOnFile(session) +:
         new CarbonIUDAnalysisRule(sparkSession) +:
         new CarbonPreInsertionCasts(sparkSession) +: customResolutionRules
+
       override val extendedCheckRules: Seq[LogicalPlan => Unit] =
-        PreWriteCheck :: HiveOnlyCheck :: Nil
+      PreWriteCheck :: HiveOnlyCheck :: Nil
+
       override val postHocResolutionRules: Seq[Rule[LogicalPlan]] =
+        new DetermineTableStats(session) +:
+        RelationConversions(conf, catalog) +:
         PreprocessTableCreation(session) +:
         PreprocessTableInsertion(conf) +:
         DataSourceAnalysis(conf) +:
+        HiveAnalysis +:
         customPostHocResolutionRules
     }
   )
-  override protected def newBuilder: NewBuilder = new CarbonInMemorySessionStateBuilder(_, _)
-}
 
+  override protected def newBuilder: NewBuilder = new CarbonSessionStateBuilder(_, _)
+}
