+++ integration/spark2/src/main/spark2.2/org/apache/spark/sql/hive/CarbonSessionUtil.scala
@@ -18,12 +18,13 @@
 package org.apache.spark.sql.hive
 
 import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.catalyst.catalog.{CatalogTable, CatalogTablePartition, ExternalCatalogUtils}
-import org.apache.spark.sql.catalyst.expressions.Expression
+import org.apache.spark.sql.catalyst.catalog.{CatalogTable, CatalogTablePartition}
 import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, SubqueryAlias}
 import org.apache.spark.sql.execution.datasources.LogicalRelation
 import org.apache.spark.sql.{CarbonDatasourceHadoopRelation, CarbonEnv, SparkSession}
 import org.apache.spark.util.CarbonReflectionUtils
+import org.apache.spark.sql.catalyst.catalog.ExternalCatalogUtils
+import org.apache.spark.sql.catalyst.expressions.Expression
 
 import org.apache.carbondata.common.logging.LogServiceFactory
 
@@ -48,10 +49,10 @@ object CarbonSessionUtil {
     var isRelationRefreshed = false
     rtnRelation match {
       case SubqueryAlias(_,
-      MatchLogicalRelation(_: CarbonDatasourceHadoopRelation, _, _)
+      LogicalRelation(_: CarbonDatasourceHadoopRelation, _, _)
       ) =>
         isRelationRefreshed = CarbonEnv.refreshRelationFromCache(name)(sparkSession)
-      case MatchLogicalRelation(_: CarbonDatasourceHadoopRelation, _, _) =>
+      case LogicalRelation(_: CarbonDatasourceHadoopRelation, _, _) =>
         isRelationRefreshed = CarbonEnv.refreshRelationFromCache(name)(sparkSession)
       case SubqueryAlias(_, relation) if
       relation.getClass.getName.equals("org.apache.spark.sql.catalyst.catalog.CatalogRelation") ||
