+++ integration/spark2/src/test/scala/org/apache/spark/carbondata/bucketing/TableBucketingTestCase.scala
@@ -18,7 +18,7 @@
 package org.apache.spark.carbondata.bucketing
 
 import org.apache.spark.sql.common.util.Spark2QueryTest
-import org.apache.spark.sql.execution.exchange.Exchange
+import org.apache.spark.sql.execution.exchange.ShuffleExchange
 import org.scalatest.BeforeAndAfterAll
 
 import org.apache.carbondata.common.exceptions.sql.MalformedCarbonCommandException
@@ -127,11 +127,7 @@ class TableBucketingTestCase extends Spark2QueryTest with BeforeAndAfterAll {
       """.stripMargin).queryExecution.executedPlan
     var shuffleExists = false
     plan.collect {
-      case s: Exchange if (s.getClass.getName.equals
-      ("org.apache.spark.sql.execution.exchange.ShuffleExchange") ||
-        s.getClass.getName.equals
-        ("org.apache.spark.sql.execution.exchange.ShuffleExchangeExec"))
-      => shuffleExists = true
+      case s: ShuffleExchange => shuffleExists = true
     }
     assert(shuffleExists, "shuffle should exist on non bucket tables")
   }
@@ -149,11 +145,7 @@ class TableBucketingTestCase extends Spark2QueryTest with BeforeAndAfterAll {
       """.stripMargin).queryExecution.executedPlan
     var shuffleExists = false
     plan.collect {
-      case s: Exchange if (s.getClass.getName.equals
-      ("org.apache.spark.sql.execution.exchange.ShuffleExchange") ||
-        s.getClass.getName.equals
-        ("org.apache.spark.sql.execution.exchange.ShuffleExchangeExec"))
-      => shuffleExists = true
+      case s: ShuffleExchange => shuffleExists = true
     }
     assert(!shuffleExists, "shuffle should not exist on bucket tables")
   }
@@ -178,11 +170,7 @@ class TableBucketingTestCase extends Spark2QueryTest with BeforeAndAfterAll {
       """.stripMargin).queryExecution.executedPlan
     var shuffleExists = false
     plan.collect {
-      case s: Exchange if (s.getClass.getName.equals
-      ("org.apache.spark.sql.execution.exchange.ShuffleExchange") ||
-        s.getClass.getName.equals
-        ("org.apache.spark.sql.execution.exchange.ShuffleExchangeExec"))
-      => shuffleExists = true
+      case s: ShuffleExchange => shuffleExists = true
     }
     assert(!shuffleExists, "shuffle should not exist on bucket tables")
     sql("DROP TABLE bucketed_parquet_table")
@@ -207,11 +195,7 @@ class TableBucketingTestCase extends Spark2QueryTest with BeforeAndAfterAll {
       """.stripMargin).queryExecution.executedPlan
     var shuffleExists = false
     plan.collect {
-      case s: Exchange if (s.getClass.getName.equals
-      ("org.apache.spark.sql.execution.exchange.ShuffleExchange") ||
-        s.getClass.getName.equals
-        ("org.apache.spark.sql.execution.exchange.ShuffleExchangeExec"))
-      => shuffleExists = true
+      case s: ShuffleExchange => shuffleExists = true
     }
     assert(shuffleExists, "shuffle should exist on non bucket tables")
     sql("DROP TABLE parquet_table")
